LPD-38508 Fix flaky check for `Liferay.SPA`
